+++ CHANGES.md
@@ -5,6 +5,7 @@
   * CHANGED: move root DNS server config to the ini file
   * CHANGED: support "ping is expected to fail" per root server and AF
   * CHANGED: all defined open DNS resolvers are now in one check
+  * NEW: added "select all checks" to GUI
 
 ### 0.7.20260311
   * NEW check: SSH host key checks